Custom traits are now in local_traits.py
