+++ simphony_ui/ui.py
@@ -1,37 +1,8 @@
 from traits.api import (HasStrictTraits, Int, Float, Enum, Instance,
-                        Str, Directory, BaseInt, BaseFloat, Array, File)
+                        Str, Directory, Array, File)
 from traitsui.api import View, Item, UItem, Tabbed, VGroup, HGroup
 import os
-
-
-class PositiveFloat(BaseFloat):
-    """ A custom trait representing a positive float """
-
-    default_value = 1.0
-
-    info_text = 'a positive float'
-
-    def validate(self, object, name, value):
-        value = super(PositiveFloat, self).validate(object, name, value)
-        if value > 0.0:
-            return value
-
-        self.error(object, name, value)
-
-
-class PositiveInt(BaseInt):
-    """ A custom trait representing a positive integer """
-
-    default_value = 1
-
-    info_text = 'a positive integer'
-
-    def validate(self, object, name, value):
-        value = super(PositiveInt, self).validate(object, name, value)
-        if value > 0:
-            return value
-
-        self.error(object, name, value)
+from simphony_ui.tests.local_traits import PositiveFloat, PositiveInt
 
 
 class GlobalParametersModel(HasStrictTraits):
